+++ java-client-sdk/lib/src/main/java/org/openapis/openapi/ResponseBodies.java
@@ -374,4 +374,42 @@ public class ResponseBodies {
 
         return res;
     }
+
+    public org.openapis.openapi.models.operations.ResponseBodyZeroValueComplexTypePtrsPostResponse responseBodyZeroValueComplexTypePtrsPost(org.openapis.openapi.models.shared.ObjWithZeroValueComplexTypePtrs request) throws Exception {
+        String baseUrl = org.openapis.openapi.utils.Utils.templateUrl(this.sdkConfiguration.serverUrl, this.sdkConfiguration.getServerVariableDefaults());
+        String url = org.openapis.openapi.utils.Utils.generateURL(baseUrl, "/anything/responseBodies/zeroValueComplexTypePtrs");
+        
+        HTTPRequest req = new HTTPRequest();
+        req.setMethod("POST");
+        req.setURL(url);
+        SerializedBody serializedRequestBody = org.openapis.openapi.utils.Utils.serializeRequestBody(request, "request", "json");
+        if (serializedRequestBody == null) {
+            throw new Exception("Request body is required");
+        }
+        req.setBody(serializedRequestBody);
+
+        req.addHeader("Accept", "application/json");
+        req.addHeader("x-speakeasy-user-agent", String.format("speakeasy-sdk/%s %s %s %s", this.sdkConfiguration.language, this.sdkConfiguration.sdkVersion, this.sdkConfiguration.genVersion, this.sdkConfiguration.openapiDocVersion));
+        
+        HTTPClient client = this.sdkConfiguration.securityClient;
+        
+        HttpResponse<byte[]> httpRes = client.send(req);
+
+        String contentType = httpRes.headers().firstValue("Content-Type").orElse("application/octet-stream");
+
+        org.openapis.openapi.models.operations.ResponseBodyZeroValueComplexTypePtrsPostResponse res = new org.openapis.openapi.models.operations.ResponseBodyZeroValueComplexTypePtrsPostResponse(contentType, httpRes.statusCode()) {{
+            responseBodyZeroValueComplexTypePtrsPost200ApplicationJSONObject = null;
+        }};
+        res.rawResponse = httpRes;
+        
+        if (httpRes.statusCode() == 200) {
+            if (org.openapis.openapi.utils.Utils.matchContentType(contentType, "application/json")) {
+                ObjectMapper mapper = JSON.getMapper();
+                org.openapis.openapi.models.operations.ResponseBodyZeroValueComplexTypePtrsPost200ApplicationJSON out = mapper.readValue(new String(httpRes.body(), StandardCharsets.UTF_8), org.openapis.openapi.models.operations.ResponseBodyZeroValueComplexTypePtrsPost200ApplicationJSON.class);
+                res.responseBodyZeroValueComplexTypePtrsPost200ApplicationJSONObject = out;
+            }
+        }
+
+        return res;
+    }
 }