+++ ruby-client-sdk/lib/sdk/responsebodies.rb
@@ -305,5 +305,44 @@ module OpenApiSDK
       end
       res
     end
+
+    sig { params(request: Shared::ObjWithZeroValueComplexTypePtrs).returns(Utils::FieldAugmented) }
+    def response_body_zero_value_complex_type_ptrs_post(request)
+
+      url, params = @sdk_configuration.get_server_details
+      base_url = Utils.template_url(url, params)
+      url = "#{base_url}/anything/responseBodies/zeroValueComplexTypePtrs"
+      headers = {}
+      req_content_type, data, form = Utils.serialize_request_body(request, :request, :json)
+      headers['content-type'] = req_content_type
+      raise StandardError, 'request body is required' if data.nil? && form.nil?
+      headers['Accept'] = 'application/json'
+      headers['x-speakeasy-user-agent'] = "speakeasy-sdk/#{@sdk_configuration.language} #{@sdk_configuration.sdk_version} #{@sdk_configuration.gen_version} #{@sdk_configuration.openapi_doc_version}"
+
+      r = @sdk_configuration.client.post(url) do |req|
+        req.headers = headers
+        Utils.configure_request_security(req, @sdk_configuration.security) if !@sdk_configuration.nil? && !@sdk_configuration.security.nil?
+        if form
+          req.body = Utils.encode_form(form)
+        elsif Utils.match_content_type(req_content_type, 'application/x-www-form-urlencoded')
+          req.body = URI.encode_www_form(data)
+        else
+          req.body = data
+        end
+      end
+
+      content_type = r.headers.fetch('Content-Type', 'application/octet-stream')
+
+      res = Operations::ResponseBodyZeroValueComplexTypePtrsPostResponse.new(
+        status_code: r.status, content_type: content_type, raw_response: r
+      )
+      if r.status == 200
+        if Utils.match_content_type(content_type, 'application/json')
+          out = Utils.unmarshal_complex(r.env.response_body, Operations::ResponseBodyZeroValueComplexTypePtrsPost200ApplicationJSON)
+          res.response_body_zero_value_complex_type_ptrs_post_200_application_json_object = out
+        end
+      end
+      res
+    end
   end
 end